Fix missing comma in SConstruct test
The comma was omitted to discourage putting anything after the last
element.  This was fine when there were multiple elements, but with only
one value, the missing comma causes this not to be a tuple as intended.
Add a comma to fix the test.  Add some additional comments on supported
C++ versions by gcc major version.