Fix TypeError in `default_opengles`, `default_sdl2`
Commit 452d58a0b5fb switched from passing a list of strings as the
right-hand side of `in` to passing a parenthesized string.  The new form
causes a TypeError if the left hand side is None:

```
>>> None in ('1', '2')
False
>>> None in ('1')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: 'in <string>' requires string as left operand, not NoneType
```

Add a trailing comma to turn the right hand side back into a list of
strings, as it was before the Raspberry Pi commit.

Fixes: 452d58a0b5fbac0cce8c40c06d19afd15e5d0dcd ("Update Raspberry Pi Mesa build for Buster & Pi 4B")