Add release note about savegame format change
zicodxx reported in issue #449 that savegames created in a new build,
then loaded in an older one, may behave incorrectly in the old build.
This cannot be fixed in the new build without reverting the logic change
which caused it.  Support for old builds has always been best effort, so
this will not be fixed.  Add a release note about it.

Reported-by: zicodxx <https://github.com/dxx-rebirth/dxx-rebirth/issues/449>