Revert "fix invalid polymodels if words_need_alignment=1"
This reverts commit 60edaa7dcefaaa92777b05233ac35145ac18e585.

This commit was brought in through a merge to obtain a different change,
and was supposed to be suppressed in preference to the change from
dfcb0fcdc.  Revert it here, since reloading the variable suffices to fix
the problem, and this commit forces a reload on all platforms, rather
than only reloading on platforms that need it.